Moved package to rnxmpp
