+++ android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="ae.teletronics.react_native_xmpp">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="rnxmpp">
 
 </manifest>
